Location to new file
