+++ Unit.h
@@ -3,6 +3,8 @@
 #include <string>
 #include "Damage.h"
 
+#include "Location.h"
+
 // By default these are set to 0 to "disable" the feature
 #define LEVEL_SIZE 0
 #define HEALTH_BONUS 0
@@ -35,45 +37,6 @@
 
 class Unit
 {
-public:
-	/**
-	 * Location object.
-	 */
-	class Location
-	{
-	public:
-		/**
-		 * The X coordinate of the location.
-		 */
-		int x;
-
-		/**
-		 * The Y coordinate of the location.
-		 */
-		int y;
-
-		/**
-		 * Constructor of the location.
-		 */
-		Location(){};
-
-		/**
-		 * Constructor of the location.
-		 * @param x The X coordinate
-		 * @param y The Y coordinate
-		 */
-		Location(int x, int y) : x(x), y(y){};
-
-		/**
-		 * The equal operator.
-		 * @param location The other Location object
-		 */
-		bool operator==(const Location &location)
-		{
-			return x == location.x && y == location.y;
-		}
-	};
-
 private:
     /**
      * The name of the character.
@@ -141,11 +104,11 @@ private:
      */
     double cooldownMultiplier = COOLDOWN_MULTIPLIER;
 
-	/**
+    /**
      *  Location of the Unit.
 	 *	Contains X and Y coordinates with 0 value by default.
      */
-	Location location;
+    Location location;
 
     /**
      * Defense abality size.
@@ -244,12 +207,12 @@ public:
      */
     Unit(const std::string &name, unsigned int maxHp, Damage damage, double attackCooldown, unsigned int xpPerLevel, unsigned int healthBonusPerLevel, unsigned int damageBonusPerLevel, unsigned int magicalDamageBonusPerLevel, double cooldownMultiplier, unsigned int defense, unsigned int defenseBonusPerLevel) : name(name), maxHp(maxHp), hp(maxHp), damage(damage), attackCooldown(attackCooldown), nextAttack(attackCooldown), xpPerLevel(xpPerLevel), healthBonusPerLevel(healthBonusPerLevel), damageBonusPerLevel(damageBonusPerLevel), magicalDamageBonusPerLevel(magicalDamageBonusPerLevel), cooldownMultiplier(cooldownMultiplier), defense(defense), defenseBonusPerLevel(defenseBonusPerLevel){};
 
-	/**
+    /**
 	 * Unit destructor.
 	 */
-	virtual ~Unit(){};
+    virtual ~Unit(){};
 
-	/**
+    /**
      * It parse a JSON object (from a JSON file) to a Unit instance.
      * Does not matter the order of the object properties, but it have to contain all required propertiy.
      * @param filename Json file's path.
@@ -343,59 +306,59 @@ public:
      */
     unsigned int getDefense() const { return this->defense; };
 
-	/**
+    /**
 	 * Get the X coordinate of the Unit.
 	 */
-	int getX() const { return location.x; }
+    int getX() const { return location.x; }
 
-	/**
+    /**
 	 * Get the Y coordinate of the Unit.
 	 */
-	int getY() const { return location.y; }
+    int getY() const { return location.y; }
 
-	/**
+    /**
 	 * Set the X coordinate of the Unit.
 	 * @param x The new X coordinate
 	 */
-	Unit &setX(const int x)
-	{
-		location.x = x;
-		return *this;
-	}
+    Unit &setX(const int x)
+    {
+        location.x = x;
+        return *this;
+    }
 
-	/**
+    /**
 	 * Set the Y coordinate of the Unit.
 	 * @param x The new Y coordinate
 	 */
-	Unit &setY(const int y)
-	{
-		location.y = y;
-		return *this;
-	}
+    Unit &setY(const int y)
+    {
+        location.y = y;
+        return *this;
+    }
 
-	/**
+    /**
 	 * Get the location of the Unit.
 	 */
-	Location getLocation() { return location; }
+    Location getLocation() { return location; }
 
-	/**
+    /**
 	 * Set the location coordinate of the Unit.
 	 * @param location The new location
 	 */
-	Unit &setLocation(Location location)
-	{
-		this->location = location;
-		return *this;
-	}
+    Unit &setLocation(Location location)
+    {
+        this->location = location;
+        return *this;
+    }
 
-	/**
+    /**
 	 * Set the location coordinate of the Unit.
 	 * @param location The new location
 	 */
-	Unit &setLocation(int x, int y)
-	{
-		this->location.x = x;
-		this->location.y = y;
-		return *this;
-	}
+    Unit &setLocation(int x, int y)
+    {
+        this->location.x = x;
+        this->location.y = y;
+        return *this;
+    }
 };
